+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added 13 trigonometry functions: SEC, CSC, SINH, COSH, TANH, COTH, SECH, CSCH, ACOT, ASINH, ACOSH, ATANH, ACOTH. (#485)
 - Added 6 engineering functions: OCT2BIN, OCT2DEC, OCT2HEX, HEX2BIN, HEX2OCT, HEX2DEC. (#497)
 - Added a configuration option to evaluate reference to an empty cells as a zero. (#476)
+- Added new error type: missing licence. (#306)
 - Added detailed error messages for error values. (#506)
 - Added ability to handle more characters in quoted sheet names. (#509)
 - Added support for escaping apostrophe character in quoted sheet names. (#64)